Fix bugs: properly check if an extension/curve/cipherSuit is GREASE.
The current version does not properly check GREASE in many cases. In
particular, it compares an extension ID, a curve ID, or a CipherSuit
against a constant value GREASE_PLACEHOLDER (0x0a0a), without
unGREASEing it first.

To fix this problem, we simply use the existing function:
isGREASEUint16() with some necessary casting.

For the records, some error messages related to this bug are:

* "applying generated spec failed: uTLS does not support 0xBABA as max version"

* "applying generated spec failed: unsupported Curve in
KeyShareExtension: CurveID(56026).To mimic it, fill the Data(key)
field manually"